Address CI issues caused by the test validators. Re-organize the validator's packages to be encapsulated within the tests' package.
